+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v1.2.0 (2022-11-05)
+### Feature
+* Add new login job for leetcode
+### Change
+* Move housekeeping to log service
+
 ## v1.1.0 (2022-02-26)
 ### Change
 * Update mastercard api link